requirements.txt
flask>=3.0
numpy
pandas
Pillow
torch
torchvision
transformers
faiss-cpu
